+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 All notable changes to the "emacs-mcx" extension will be documented in this file.
 
 ## [Unreleased]
-
 ### Add
 - `emacs-mcx.enableDigitArgument` config, #1256.
+- Regexp search and replace commands, #1257.
 
 ## [0.40.1] - 2022-03-20
 ### Fix